fix(launch-green): green the full test suite for the paid-launch branch
Test-suite fixes (full suite now 894 passed / 0 failed / 4 skipped):
- AI discovery: add the Ai* component schemas to contracts/openapi.v1.json
  (AiDiscoveryContractTests were missing every Ai* schema).
- Contract drift: remove the /sse and /tracks/{id}/readiness redirect aliases
  from the contract (minimal-API redirects, not controller actions; 307/308) and
  mark the /sse map ExcludeFromDescription so it is not re-emitted.
- AI search: cast TrendingScore to double in the "trending" sort so the ORDER BY
  is provider-portable (SQLite cannot ORDER BY decimal -> was a 500); + regression test.
- Pricing contract-truth: drop the retired exclusive/copyright-buyout assertions
  from the 3 stale TrackDetail tests. Those fields were intentionally removed from
  the public track DTO (licensing model retired; enforced by
  LicensingLeakRegressionTests + project rule); re-adding them was not an option.
- Cover art: update the stale Upload_WithCoverArt test to assert the bare object
  key the WIP deliberately switched to (bare "covers/..." resolves via the /images
  proxy; the old "/uploads/covers/..." form 404'd).

streamAvailable + credit resilience (with tests):
- AI single-track detail/preview report Available from a real object-existence
  probe, so missing-audio tracks are not advertised as playable.
- ReleaseCreditService.TryChargeAsync retries Postgres serialization conflicts
  (40001/40P01) so a concurrent last-credit charge returns cleanly, not a 500;
  idempotency is re-checked on retry (no double-charge).

Also folds in a concurrent session's subscription-session-idempotency work present
in the working tree (AddSubscriptionSessionIdempotency migration, StripeWebhookService
+ Subscription changes, SubscriptionGrantWebhookTests) -- not authored here but green
and committed together to keep the branch consistent.

+++ src/Cambrian.Persistence/Repositories/TrackSorting.cs
@@ -16,7 +16,10 @@ public static class TrackSorting
             "pricelowtohigh" or "price_asc" or "price" => query.OrderBy(t => t.Price),
             "pricehightolow" or "price_desc" => query.OrderByDescending(t => t.Price),
             "newest" or "recent" or "created" => query.OrderByDescending(t => t.CreatedAt),
-            "trending" or "popular" => query.OrderByDescending(t => t.TrendingScore)
+            // Cast the decimal score to double so the ORDER BY is provider-portable: SQLite
+            // throws on ORDER BY a decimal expression but can order CAST(... AS REAL); PostgreSQL
+            // handles both. Ordering is otherwise identical (monotonic cast).
+            "trending" or "popular" => query.OrderByDescending(t => (double)t.TrendingScore)
                                             .ThenByDescending(t => t.CreatedAt),
             "title" => query.OrderBy(t => t.Title),
             _ => query.OrderByDescending(t => t.CreatedAt)
